+++ klib/usb/device/dfu.hpp
@@ -64,7 +64,7 @@ namespace klib::usb::device {
 
         // device descriptor for the dfu device
         const __attribute__((aligned(4))) static inline descriptor::device device = {
-            .bcdUSB = static_cast<uint16_t>(setup::usb_version::usb_v1_1),
+            .bcdUSB = setup::usb_version::usb_v1_1,
             .bDeviceClass = descriptor::class_type::use_class_info,
             .bDeviceSubClass = 0x00,
             .bDeviceProtocol = 0x00,
