+++ klib/usb/device/mass_storage.hpp
@@ -55,7 +55,7 @@ namespace klib::usb::device {
 
         // device descriptor for the mass storage class
         const __attribute__((aligned(4))) static inline descriptor::device device = {
-            .bcdUSB = static_cast<uint16_t>(setup::usb_version::usb_v2_0),
+            .bcdUSB = setup::usb_version::usb_v2_0,
             .bDeviceClass = descriptor::class_type::use_class_info,
             .bDeviceSubClass = 0x00,
             .bDeviceProtocol = 0x00,
@@ -104,7 +104,7 @@ namespace klib::usb::device {
 
         // device qualifier
         const __attribute__((aligned(4))) static inline descriptor::qualifier qualifier = {
-            .bcdUSB = static_cast<uint16_t>(setup::usb_version::usb_v2_0),
+            .bcdUSB = setup::usb_version::usb_v2_0,
             .bDeviceClass = descriptor::class_type::use_class_info,
             .bDeviceSubClass = 0x00,
             .bDeviceProtocol = 0x00,
